v1.0.0 1. the Eye program run in background to protect my eye. it use a config file like : [General] time_unit = min work_time = 55 rest_time = 5 image_file = rei.jpg
the config file means : you need to take a rest for more than 5 minutes
every time you work for 55 minute. When you need to take a rest, the
screen will be coved by a image which is from image_file(rei.jpg)

the program has a tray, and you can use the tray to config, quit, and
make a imediate rest.
2. the program can only work in Linux now.